+++ src/commands/cmd_key.rs
@@ -91,6 +91,7 @@ fn run_list(global_args: &GlobalArgs) -> Result<()> {
         ssh_pubkey: global_args.ssh_pubkey.clone(),
         ssh_privatekey: global_args.ssh_privatekey.clone(),
         dry_backend: false,
+        cached: !global_args.no_cache,
     })?;
 
     let mut table = Table::new();
@@ -121,6 +122,7 @@ fn run_add(global_args: &GlobalArgs, args: &AddArgs) -> Result<()> {
         ssh_pubkey: global_args.ssh_pubkey.clone(),
         ssh_privatekey: global_args.ssh_privatekey.clone(),
         dry_backend: false,
+        cached: !global_args.no_cache,
     })?;
 
     let key_manager = KeyManager::new(backend.clone());
@@ -157,6 +159,7 @@ fn run_delete(global_args: &GlobalArgs, args: &DeleteArgs) -> Result<()> {
         ssh_pubkey: global_args.ssh_pubkey.clone(),
         ssh_privatekey: global_args.ssh_privatekey.clone(),
         dry_backend: false,
+        cached: !global_args.no_cache,
     })?;
     let key_manager = KeyManager::new(backend.clone());
     let (_id, path) = key_manager.find_id_with_prefix(&args.id)?;
@@ -170,6 +173,7 @@ fn run_password_change(global_args: &GlobalArgs, _args: &PasswordChangeArgs) -> 
         ssh_pubkey: global_args.ssh_pubkey.clone(),
         ssh_privatekey: global_args.ssh_privatekey.clone(),
         dry_backend: false,
+        cached: !global_args.no_cache,
     })?;
 
     let key_manager = KeyManager::new(backend.clone());