+++ src/commands/cmd_sync.rs
@@ -59,6 +59,7 @@ pub fn run(global_args: &GlobalArgs, args: &CmdArgs) -> Result<()> {
         ssh_pubkey: global_args.ssh_pubkey.clone(),
         ssh_privatekey: global_args.ssh_privatekey.clone(),
         dry_backend: false,
+        cached: false,
     })?;
 
     let (_repo, _, lock_handle) = Repository::try_open_with_lock(
@@ -74,6 +75,7 @@ pub fn run(global_args: &GlobalArgs, args: &CmdArgs) -> Result<()> {
         ssh_pubkey: args.dst_ssh_pubkey.clone(),
         ssh_privatekey: args.dst_ssh_privatekey.clone(),
         dry_backend: false,
+        cached: false,
     })?;
     dst_backend.create()?; // Create the backend to create the directory if it doesn't exist.