+++ tests/integration_tests/test_cmd_mount.rs
@@ -162,6 +162,7 @@ mod tests {
             ssh_pubkey: None,
             ssh_privatekey: None,
             pack_size_mib: DEFAULT_DEFAULT_PACK_SIZE_MIB,
+            no_cache: true,
         };
         set_global_opts_with_args(&global);
 
@@ -289,6 +290,7 @@ mod tests {
             ssh_pubkey: None,
             ssh_privatekey: None,
             pack_size_mib: DEFAULT_DEFAULT_PACK_SIZE_MIB,
+            no_cache: true,
         };
         set_global_opts_with_args(&global);